Dyadic

Description

This module contains a purely Haskell implementation of dyadic rationals, suitable for interval arithmetic. A faster implementation of dyadic rationals would use a fast arbitrary-precision floating-point library, such as MPFR and the related hmpfr Haskell bindings for it.

A dyadic number is a rational whose denominator is a power of 2. We also include positive and negative infinity, these are useful for representing infinite intervals. The special value NaN (not a number) is included as well in order to follow more closely the usual floating-point format, but is not used in the interval computations because there we use [-inf, +inf] to represent the completely undefined value.

Documentation

class (Show q, Ord q) => ApproximateField q where

An approximate field is a structure in which we can perform approximate arithmetical operations. The typical example is the ring of dyadic rational numbers: division of dyadic rationals is only approximate, and even though the other operations (+, -, *) can be peformed exactly, it is too expensive and unecessary to do so in an interval computation. Therefore, we want approximate versions of all operations.

The approximate operations take a Stage argument which tells whether the result of the operation should be rounded up or down, in the sense of the linear ordering of the structure, and how precise the result should be.

(Missing explanation of what exactly an approximate field is supposed to be.)

Minimal complete definition

normalize, size, log2, midpoint, zero, positive_inf, negative_inf, toFloat, toRational', app_add, app_sub, app_mul, app_inv, app_div, app_negate, app_abs, app_signum, app_fromInteger, app_fromInteger_ws, app_shift

Methods

normalize :: Stage -> q -> q

size

Arguments

:: q

-> Int
the size of the number (memory usage)

log2

Arguments

:: q

-> Int
log2 q is a number k such that 2^k <= abs q <= 2^(k+1).

midpoint

Arguments

:: q

-> q

-> q
exact midpoint

zero :: q

positive_inf :: q

negative_inf :: q

toFloat :: q -> Double

toRational' :: q -> Rational

app_add :: Stage -> q -> q -> q

app_sub :: Stage -> q -> q -> q

app_mul :: Stage -> q -> q -> q

app_inv :: Stage -> q -> q

app_div :: Stage -> q -> q -> q

app_negate :: Stage -> q -> q

app_abs :: Stage -> q -> q

app_signum :: Stage -> q -> q

app_fromInteger :: Stage -> Integer -> q

app_fromInteger_ws :: Integer -> q

app_shift

Arguments

:: Stage

-> q

-> Int

-> q
shift by a power of 2

app_fromRational :: Stage -> Rational -> q

data Dyadic

A dyadic number is of the form m * 2^e where m is the mantissa and e is the exponent.

Constructors

Dyadic

Fields

mant :: Integer

expo :: Int

PositiveInfinity

NegativeInfinity

NaN
not a number, result of undefined arithmetical operation