Interval

Description

This module defines the interval domain, i.e., the space of intervals. Actually, what we define as a base for such a domain because our intervals have rational endpoints (to be exact, the endpoints are elements of an approximate field). The actual interval domains is defined in the module Reals.

Documentation

data Interval q

An interval is represented by a lower and upper endpoint. We do not require that the lower endpoint be smaller or equal to the upper one. In other words, we allow the usual as well as back-to-front intervals. This is useful in certain kinds of computations.

A possible optimization: represent an interval with its center point and radius, where the radius is not a precise number (has a small mantissa). This can save up to 50% of space, but it is not clear how to treat back-to-front intervals then. Presumably with negative radii, except I have never worked out how to implement interval multiplication then.

Constructors

Interval

Fields

lower :: q

upper :: q

class ApproximateField q => IntervalDomain q where

Minimal complete definition

width

Methods

iless :: Interval q -> Interval q -> Bool

imore :: Interval q -> Interval q -> Bool

iadd :: Stage -> Interval q -> Interval q -> Interval q

isub :: Stage -> Interval q -> Interval q -> Interval q

imul :: Stage -> Interval q -> Interval q -> Interval q

iinv :: Stage -> Interval q -> Interval q

idiv :: Stage -> Interval q -> Interval q -> Interval q

iabs :: Stage -> Interval q -> Interval q

inormalize :: Stage -> Interval q -> Interval q

embed :: Stage -> q -> Interval q

split :: Interval q -> (Interval q, Interval q)

width :: Interval q -> q