Reals

Description

We implement real numbers as the completion of dyadic intervals. The whole construction is parametrized by an approximate field, an example of which is Dyadic.

Documentation

type RealNum q = Staged (Interval q)

A real number is implemented as a staged dyadic interval Interval q where q is the underlying approximate field (in practiec these are dyadic rationals). RealNum q can be used to represent not only real numbers but also the elements of the interval domain, including the back-to-front intervals.

newtype ClosedInterval q

The value ClosedInterval(a,b) represents the closed interval [a,b] as a subspace of the reals.

Constructors

ClosedInterval (q, q)

exact :: RealNum Dyadic -> RealNum Dyadic

This is a convenience function which allows us to write exact 1.3 as a conversion from floating points to real numbers. There probably is a better way of doing this.

toReal :: IntervalDomain q => q -> RealNum q

Function convert elements of type q in elements of type RealNum q.

lim :: IntervalDomain q => (Int -> (RealNum q, q)) -> RealNum q

Reals form a complete space, which means that every Cauchy sequence of reals has a limit. In the implementation this is manifested by the existence of an operator which computes the limit of a Cauchy sequence. The error bounds for the sequence are given explicitly.

approx_to :: IntervalDomain q => RealNum q -> Int -> RoundingMode -> (q, Int)

Reals form an Archimedean field. Topologically speaking, this means that the underlying approximate field q is dense in the reals. Computationally this means that we can compute arbitrarily good q-approximations to real numbers. The function 'approx_to x k r' computes an approximation a of type q which is within 2^-k of x.

fac :: Rational -> Rational

tI :: Rational -> Integer

ilogb :: Integer -> Integer -> Int