Space

Description

Basic definitions od spaces and their properties

Documentation

type Sigma = Staged Bool

The Sierpinski space Sigma is represented by staged booleans

sor :: Sigma -> Sigma -> Sigma

Disjunction for Sierpinski space

sand :: Sigma -> Sigma -> Sigma

Conjunction for Sierpinski space

force :: Sigma -> Bool

Force a value in the Sierpinski space into Booleans. This may diverge as bottom cannot be reliably detected.

class Hausdorff t where

A space is Hausdorff if inequality, here called apart, is an open relation.

Methods

apart :: t -> t -> Sigma

class Discrete t where

A space is Discrete if equality, here called equal, is an open relation.

Methods

equal :: t -> t -> Sigma

class Compact s t | s -> t where

Suppose the type s represents a family of subspaces of t. The typical example is that t is the type of reals and s is the type of closed intervals. Then the subspaces represented by s are compact subspaces of t if the universal quantifier is a continuous map from t -> Sigma to Sigma.

Methods

forall :: s -> (t -> Sigma) -> Sigma

class Overt s t | s -> t where

Suppose the type s represents a family of subspaces of t. The typical example is that t is the type of reals and s is the type of closed intervals. Then the subspaces represented by s are overt subspaces of t if the existential quantifier is a continuous map from t -> Sigma to Sigma.

Methods

exists :: s -> (t -> Sigma) -> Sigma

class LinearOrder t where

The real numbers are strictly linearly ordered by open relation <, we define a class that expresses that fact.

Minimal complete definition

less

Methods

less :: t -> t -> Sigma

more :: t -> t -> Sigma